Move darwinia-ethereum bridge subql graphql query to subquery-d2e crate
