--- bridges/darwinia-ethereum/bridge/src/graphql/authorities_change_signed_event.query.graphql
@@ -1,23 +1,0 @@
-query authoritiesChangeSignedPage($from: Int!, $first: Int!) {
-  authoritiesChangeSignedEvents(
-    first: $first
-    orderBy: AT_BLOCK_NUMBER_ASC
-    filter: {
-      atBlockNumber: {
-        greaterThan: $from
-      }
-    }
-  ) {
-    nodes {
-      atBlockNumber
-      term
-      newAuthorities
-      signatures {
-        nodes {
-          account,
-          relayAuthoritySignature
-        }
-      }
-    }
-  }
-}
