--- bridges/darwinia-ethereum/bridge/src/graphql/latest_schedule_mmr_root_event.query.graphql
@@ -1,13 +1,0 @@
-query ScheduleMMRRootPage {
-  scheduleMMRRootEvents(
-    first: 1
-    orderBy: AT_BLOCK_NUMBER_DESC
-  ) {
-    nodes {
-      id
-      atBlockNumber
-      eventBlockNumber
-      emitted
-    }
-  }
-}
