--- bridges/darwinia-ethereum/bridge/src/graphql/mmr_root_signed_events.query.graphql
@@ -1,23 +1,0 @@
-query MMRRootPage($from: Int!, $first: Int!) {
-  mMRRootSignedEvents(
-    first: $first
-    orderBy: AT_BLOCK_NUMBER_ASC
-    filter: {
-      atBlockNumber: {
-        greaterThan: $from
-      }
-    }
-  ) {
-    nodes {
-      atBlockNumber,
-      eventBlockNumber,
-      mmrRoot,
-      signatures {
-        nodes {
-          account,
-          relayAuthoritySignature,
-        }
-      }
-    }
-  }
-}
