--- bridges/darwinia-ethereum/bridge/src/graphql/schedule_authorities_change_event.query.graphql
@@ -1,16 +1,0 @@
-query scheduleAuthoritiesChangePage($from: Int!, $first: Int!) {
-  scheduleAuthoritiesChangeEvents(
-    first: $first
-    orderBy: AT_BLOCK_NUMBER_ASC
-    filter: {
-      atBlockNumber: {
-        greaterThan: $from
-      }
-    }
-  ) {
-    nodes {
-      atBlockNumber
-      message
-    }
-  }
-}
