+++ frame/assistants/subquery-d2e/src/graphql/authorities_change_signed_event.query.graphql
