+++ frame/assistants/subquery-d2e/src/graphql/latest_schedule_mmr_root_event.query.graphql
