+++ frame/assistants/subquery-d2e/src/graphql/mmr_root_signed_events.query.graphql
