+++ frame/assistants/subquery-d2e/src/graphql/schedule_authorities_change_event.query.graphql
